requirements.txt
praw==7.7.1
boto3==1.34.0
pandas==2.1.4
python-dotenv==1.0.0
apache-airflow==2.8.0